Minor fixes and cleanups
